farabi/bot-385/embed hotjar attribution (#9419)
* fix: added new attributes for hotjar

* refactor: separated hotjar script to a function in utils

* fix: updated attribute

* fix: added optional for client login id

* chore: added comment for attribution code
