+++ packages/bot-web-ui/src/app/app-content.jsx
@@ -6,6 +6,7 @@ import { Audio, BotNotificationMessages, Dashboard, NetworkToastPopup, RouteProm
 import BotBuilder from 'Components/dashboard/bot-builder';
 import TransactionDetailsModal from 'Components/transaction-details';
 import GTM from 'Utils/gtm';
+import hotjar from 'Utils/hotjar';
 import { useDBotStore } from 'Stores/useDBotStore';
 import BlocklyLoading from '../components/blockly-loading';
 import './app.scss';
@@ -31,24 +32,7 @@ const AppContent = observer(() => {
     }, [is_dark_mode_on]);
 
     React.useEffect(() => {
-        /**
-         * Inject: External Script Hotjar - for DBot only
-         */
-        (function (h, o, t, j) {
-            /* eslint-disable */
-            h.hj =
-                h.hj ||
-                function () {
-                    (h.hj.q = h.hj.q || []).push(arguments);
-                };
-            /* eslint-enable */
-            h._hjSettings = { hjid: 3050531, hjsv: 6 };
-            const a = o.getElementsByTagName('head')[0];
-            const r = o.createElement('script');
-            r.async = 1;
-            r.src = t + h._hjSettings.hjid + j + h._hjSettings.hjsv;
-            a.appendChild(r);
-        })(window, document, 'https://static.hotjar.com/c/hotjar-', '.js?sv=');
+        hotjar(client);
     }, []);
 
     React.useEffect(() => {
